backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.31.1
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
pydantic-settings==2.6.0
pyjwt==2.9.0
pika==1.3.2
email-validator==2.2.0
